Enable LLVM for ALL TENSIX cores including DM dispatch kernels
Major changes to enable full LLVM compilation:

1. noc_index_t type: Defined in noc_parameters.h, used consistently
   across NOC API (73 functions), dataflow API (47 callers), firmware
   files, and remote_circular_buffer.h. Eliminates uint8_t/uint32_t
   mismatch that caused GCC 15 template body checking errors.

2. Dispatch kernel fixes:
   - NOC_INDEX/NOC_MODE defaults in dataflow_api_common.h
   - NOC_INDEX/NOC_MODE added to dispatch kernel configs
   - Explicit casts at NOC template call sites in dataflow_api.h
   - -Wno-error=template-body for GCC 15 compatibility
   - -mno-relax and -Wl,--no-relax to prevent LLVM relaxation relocs

3. Relaxed relocation validator in tt_memory.cpp to accept
   LLVM-generated relocations with --emit-relocs

4. build.cpp: Enable LLVM for ALL TENSIX (firmware + kernels)

Remaining: reader/writer DM kernel linker scripts need AT() for
.data LMA (same fix as firmware linker scripts).

+++ tt_metal/impl/dispatch/kernels/cq_dispatch_subordinate.cpp
@@ -348,7 +348,7 @@ void kernel_main() {
     uint32_t total_pages_acquired = 0;
     while (!done) {
         DeviceZoneScopedN("CQ-DISPATCH-SUBORDINATE");
-        cb_acquire_pages_dispatch_s<my_noc_xy, my_dispatch_cb_sem_id>(1);
+        cb_acquire_pages_dispatch_s<my_noc_xy, my_dispatch_cb_sem_id>((uint32_t)1);
 
         volatile CQDispatchCmd tt_l1_ptr* cmd = (volatile CQDispatchCmd tt_l1_ptr*)cmd_ptr;
         DeviceTimestampedData("process_cmd_d_dispatch_subordinate", (uint32_t)cmd->base.cmd_id);
@@ -365,7 +365,7 @@ void kernel_main() {
         cmd_ptr = round_up_pow2(cmd_ptr, cb_page_size);
         // Release a single page to prefetcher. Assumption is that all dispatch_s commands fit inside a single page for
         // now.
-        cb_release_pages_dispatch_s<upstream_noc_xy, upstream_dispatch_cb_sem_id>(1);
+        cb_release_pages_dispatch_s<upstream_noc_xy, upstream_dispatch_cb_sem_id>((uint32_t)1);
         if (cmd_ptr == cb_end) {
             cmd_ptr = cb_base;
         }
